requirements.txt
anyio==3.6.1
asgiref==3.4.1
attrs==22.2.0
awsebcli==3.20.5
backports.entry-points-selectable==1.1.1
bcrypt==4.0.1
beautifulsoup4==4.10.0
bech32==1.2.0
bleach==6.0.0
blessed==1.20.0
boto3==1.17.58
botocore==1.29.81
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.3
crispy-bootstrap5==0.6
cryptography==36.0.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
Django==4.0.1
django-bootstrap5==21.3
django-ckeditor-5==0.2.4
django-crispy==1.8.1
django-crispy-forms==1.13.0
django-email-verification==0.3.1
django-environ==0.9.0
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-recaptcha==3.0.0
django-storages==1.11.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20
fastapi==0.78.0
filelock==3.4.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
image==1.5.33
importlib-metadata==6.6.0
install==1.3.5
jaraco.classes==3.2.3
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
oauthlib==3.2.2
packaging==20.9
paramiko==3.1.0
pathspec==0.10.1
piexif==1.1.3
Pillow==8.4.0
pkginfo==1.9.6
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
Pygments==2.15.1
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
PyYAML==5.4.1
qrcode==6.1
readme-renderer==37.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
s3transfer==0.4.0
segno==1.5.2
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.19.1
termcolor==1.1.0
texttable==1.6.7
twine==4.0.2
typing_extensions==4.2.0
urllib3==1.26.8
uvicorn==0.18.2
uvloop==0.16.0
virtualenv==20.10.0
watchfiles==0.15.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.3
whitenoise==6.4.0
zebedee==0.0.7
zipp==3.15.0
